requirements.txt
python-telegram-bot[webhooks]==21.3
anthropic>=0.49.0
openai>=1.40.0
openpyxl==3.1.2
httpx>=0.27.0
matplotlib
rapidfuzz>=3.0.0
# Dashboard API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
psycopg2-binary>=2.9.9
cloudinary
PyJWT>=2.8.0
apscheduler>=3.10.0

sentry-sdk[fastapi]>=2.0.0
prometheus-client>=0.20.0